dfgdfgdfg

+++ MayhemBTH2017/MayhemBTH2017/Mesh.cpp
@@ -43,23 +43,24 @@ bool Mesh::LoadMesh(Vertex * vertices, uint64_t numVerts)
 	glBindBuffer(GL_ARRAY_BUFFER, m_buffer);
 
 	// Copy data to the gpu.
-	glBufferData(GL_ARRAY_BUFFER, sizeof (Vertex) * numVerts, vertices, GL_STATIC_DRAW);
-	
-	GLuint offset = 0;
+	glBufferData(GL_ARRAY_BUFFER, sizeof(Vertex) * numVerts, vertices, GL_STATIC_DRAW);
+
+	uint64_t offset = 0;
 
 	// Position.
 	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(Vertex), BUFFER_OFFSET(offset));
 	offset += sizeof(glm::vec3);
-	
+
 	// Normal.
 	glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, sizeof(Vertex), BUFFER_OFFSET(offset));
 	offset += sizeof(glm::vec3);
-	
+
 	// Texture Coordinates.
 	glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, sizeof(Vertex), BUFFER_OFFSET(offset));
 
 	// Unbind
 	glBindVertexArray(0);
+
 
 	m_isLoaded = true;
 	return true;
